Created script for sound bars, but encountered an issue with accessing AudioSampleCollector script
I suspect the error is due to an issue with VisualStudio on mac havng issues with Intellisense, so I'm going to try and fix it on PC later. Project does not play currently. For reference, this is not working but I'm pretty sure it should: https://answers.unity.com/questions/63922/how-do-you-access-a-script-attached-to-another-obj.html